requirements.txt
ollama>=0.6.0
python-gitlab>=4.0.0
python-dotenv>=1.0.0
urllib3>=2.0.0
